id 타입 Integer에서 Long으로 변경

+++ src/main/java/com/allrounders/goalkeeper/domain/Goal.java
@@ -18,7 +18,7 @@ import java.util.List;
 public class Goal {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer goalId;
+    private Long goalId;
 
     @Column(length = 20, nullable = false)
     private String title;
